% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICeITH.R
\name{ICeITH}
\alias{ICeITH}
\title{Estimates cell-type-specific proportions and measure intra-tumor
heterogeneity.}
\usage{
ICeITH(
  Y,
  reference,
  sampIndex,
  rho = 1,
  alpha_init = 1,
  eps = 1e-04,
  maxIters = 50
)
}
\arguments{
\item{Y}{A mixed gene expression matrix. It is a \eqn{G}
by \eqn{S} matrix, where \eqn{G} is the number of genes and \eqn{S}
is the number of samples for mixed gene expression matrix.}

\item{reference}{A list contains estimated cell-type-specific mean
expression and variability.}

\item{sampIndex}{A vector to label patient samples.
It is a \eqn{S} by 1 vector.}

\item{rho}{A scale factor. The default is 1.}

\item{alpha_init}{A hyper-parameter. The default is 1.}

\item{eps}{The convergence criterion. The default is 10^(-4).}

\item{maxIters}{The maximum number of iterations for optimization. The default
is 50.}
}
\value{
A list contains the estimated outcomes
\itemize{
    \item CT - A matrix of estimated cell-type-specific proportions.
    It is a \eqn{S} by \eqn{K} matrix.
    \item ITH - A vector of estimated intratumor heterogeneity class.
    It is a \eqn{N} by 1 vector, where \eqn{N} is the number of patients.
    \item meta.data - Metadata over iterations.
}
}
\description{
This function is designed to estimate the cell-type
specific proportions and measure intra-tumor heterogeneity from
multiple mixed samples from patients.
}
\author{
Peng Yang, Ziyi Li.
}
