% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICeITH.R
\name{refEst}
\alias{refEst}
\title{Estimates cell-type-specific mean expression and variability.}
\usage{
refEst(X, cts)
}
\arguments{
\item{X}{A cell-type-specific reference matrix. It is a \eqn{G}
by \eqn{M} matrix, where \eqn{G} is the number of genes and \eqn{M}
is the number of samples for reference.}

\item{cts}{A vector to label cell type in the reference matrix.
It is a \eqn{M} by 1 vector.}
}
\value{
mu_gk A matrix of estimated cell-type-specific mean expression.
It is a \eqn{G} by \eqn{K} matrix, where \eqn{K} is the number
of cell types of interests.

var_gk A matrix pf estimated cell-type-specific variability.
It is a \eqn{G} by \eqn{K} matrix.
}
\description{
This function is designed to estimate the cell-type
specific proportions and measure intra-tumor heterogeneity from
multiple mixed samples from patients.
}
\author{
Peng Yang, Ziyi Li.
}
