% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_func.R
\name{sim_func}
\alias{sim_func}
\title{Function to simulate multi-region gene expression data.}
\usage{
sim_func(
  K = 6,
  G = 250,
  lowS = 4,
  maxS = 8,
  N = 20,
  rho_gk = 1,
  nRef = 100,
  lowC = 0.1,
  maxC = 5
)
}
\arguments{
\item{K}{Number of cell types.}

\item{G}{Number of marker genes.}

\item{lowS}{The minimal number of samples per subject.}

\item{maxS}{The maximal number of samples per subject.}

\item{N}{Number of patients.}

\item{rho_gk}{Scale parameter to control the information borrowing
from the reference profile.}

\item{nRef}{Number of samples for the reference profile.}

\item{lowC}{Lowest intratumor heterogeneity score.}

\item{maxC}{Highest intratumor heterogeneity score.}
}
\value{
A list contains the simulated paramters
\itemize{
    \item Y - Mixed gene expression.
    \item ct_g - Cell-type-specific genes.
    \item ct_s - Cell-type-specific samples from the reference profile.
    \item I_i - Sample index to subjects.
    \item W_igk - Hidden cell-type-specific gene expression for each sample.
    \item mu_gk - True reference mean expression.
    \item mu_igk - True patient-specific mean expression.
    \item lambda_gk - True gene and cell-type-specific variability.
    \item h_sk - Cell-type-specific proportions.
    \item ITH - Patient-specific intratumor level.
    \item pi_k - Mean cell-type-specific proportions.
}
}
\description{
This function is designed to generate the cell-type-specific
reference profile and the multi-region gene expression data to perform
reference-based deconvolution by ICeITH model.
}
