% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICeITH.R
\name{update_Allpara}
\alias{update_Allpara}
\title{Update function for all parameters.}
\usage{
update_Allpara(
  logY,
  lambda_sg,
  xi_sk0,
  tau_gk0,
  gamma_igk0,
  mu_gk,
  alpha_gk,
  beta_gk,
  rho_gk,
  I_i,
  C_i0,
  pi_k0,
  eps = 1e-04,
  maxIters = 10
)
}
\arguments{
\item{logY}{log-trans input matrix.}

\item{lambda_sg}{observed variability.}

\item{xi_sk0}{initialized cell-type proportions.}

\item{tau_gk0}{initialized cell-type variability.}

\item{gamma_igk0}{initialized patient-specific mean expression.}

\item{mu_gk}{cell-type-specific mean expression.}

\item{alpha_gk}{gamma parameter for variability.}

\item{beta_gk}{gamma parameter for variability.}

\item{rho_gk}{gamma parameter for variability.}

\item{I_i}{sample index for patient}

\item{C_i0}{initialized ITH-S.}

\item{pi_k0}{nitialized mean cell-type proportions.}

\item{eps}{convergence criteria.}

\item{maxIters}{the maximum number of iterations used.}
}
\value{
res A list of estimated parameters across iterations.
}
\description{
This function is designed to estimate the cell-type
specific proportions and measure intra-tumor heterogeneity from
multiple mixed samples from patients.
}
\author{
Peng Yang, Ziyi Li.
}
